Break CLI into sub-CLIs.
