+++ indra_db/cli/preassembly.py
@@ -1,4 +1,4 @@
-from argparse import ArgumentParser
+import click
 from datetime import datetime
 
 from indra_db import get_db
@@ -6,6 +6,7 @@ from indra_db.exceptions import IndraDbException
 from indra_db.preassembly.submitter import VALID_STATEMENTS, \
     PreassemblySubmitter
 
+from .util import format_date
 
 def filter_updates(stmt_type, pa_updates):
     return {u.run_datetime for u in pa_updates if u.stmt_type == stmt_type}
@@ -75,3 +76,60 @@ def run_preassembly(mode, project_name):
     ps = PreassemblySubmitter(basename, mode, project_name=project_name)
     ps.set_max_jobs(4)
     ps.run(need_to_update, 100000, True, stagger=600, poll_interval=120)
+
+
+@click.group()
+def pa():
+    """Manage the preassembly pipeline."""
+
+
+@pa.command()
+@click.argument('task', type=click.Choice(['create', 'update']),
+                required=True)
+@click.argument('project-name', required=False)
+def run(task, project_name):
+    """Manage the indra_db preassembly.
+
+    \b
+    Tasks:
+     - "create": populate the pa_statements table for the first time (this
+       requires that the table be empty).
+     - "update": update the existing content in pa_statements with the latest
+       from raw statements.
+
+    A project name is required to tag the AWS instances with a "project" tag.
+    """
+    run_preassembly(task, project_name)
+
+
+@pa.command('list')
+@click.option('-r', '--with-raw', is_flag=True,
+              help="Include the latest datetimes for raw statements of each "
+                   "type. This will take much longer.")
+def show_list(with_raw):
+    """List the latest updates for each type of Statement."""
+    import tabulate
+
+    db = get_db('primary')
+    rows = [(st, lu) for st, lu in list_last_updates(db).items()]
+    header = ('Statement Type', 'Last Update')
+    if with_raw:
+        print("This may take a while...", end='', flush=True)
+        raw_stmt_dates = list_latest_raw_stmts(db)
+        print("\r", end='')
+        new_rows = []
+        for st, lu in rows:
+            raw_date = raw_stmt_dates.get(st)
+            if raw_date is None:
+                new_rows.append((st, format_date(lu), "[None]", "No"))
+            else:
+                new_rows.append((st, format_date(lu), format_date(raw_date),
+                                 "Yes" if raw_date > lu else "No"))
+        rows = new_rows
+        header += ('Latest Raw Stmt', 'Needs Update?')
+    else:
+        rows = [(st, format_date(lu)) for st, lu in rows]
+    rows.sort()
+    print(tabulate.tabulate(rows, header))
+
+
