+++ indra_db/cli/util.py
@@ -1,0 +1,7 @@
+from datetime import datetime
+
+
+def format_date(dt):
+    if not isinstance(dt, datetime):
+        return dt
+    return dt.strftime("%Y %b %d %I:%M%p")
